refactor(e2e): rename instance-* test files to server-*
Follow-up to the frontend Phase 7 sweep. Mechanical filename rename plus
import path updates; no test logic changes.
