+++ frontend/e2e/oauth-authorize.test.ts
@@ -4,7 +4,7 @@ import {
 	startSecondServer,
 	stopSecondServer,
 	createUserOnRemote
-} from './fixtures/multiInstance';
+} from './fixtures/multiServer';
 import type { ServerInfo } from './fixtures/server';
 import { TIMEOUTS } from './constants';
