+++ frontend/e2e/servers.test.ts
@@ -5,7 +5,7 @@ import {
 	stopSecondServer,
 	createUserOnRemote,
 	connectRemoteInstance
-} from './fixtures/multiInstance';
+} from './fixtures/multiServer';
 import type { ServerInfo } from './fixtures/server';
 import { TIMEOUTS } from './constants';
